requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
numpy==1.23.0
pandas==1.4.3
plotly==5.9.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
six==1.16.0
tenacity==8.0.1
urllib3==1.26.9
